Added mlReplyGenerator() method
private method mlReplyGenerator(..) accepts a message string and generates a reply to that string and returns that string. It uses the transformers library of python to generate a reply to each of the sentences in the scammer's email content and returns an appended string of all the replies to the sentences. nltk is used to tokenize the paragraph into sentences.